Extract home page into its own file
